ADuCM3029 BH1790GLC Example. Interrupt file was updated with the myState variable

+++ ADuCM3029/Examples_Drivers/BH1790GLC/inc/interrupts.h
@@ -38,8 +38,8 @@ extern void GP_Tmr0_Int_Handler ( void );
 
 /**@brief Variables.
  */
-extern volatile uint8_t  myState;
-extern volatile uint8_t  *myPtr;
+extern volatile uint32_t  myState;
+extern volatile uint8_t   *myPtr;
 
 
 #ifdef __cplusplus
